improved AI, minor level adjustment
characters stop moving when player approaches them, then resumes path when they leave. Bool setting in inspector to determine if the character navmesh should go in order or not. Characters look at player when in ghost form (need to change this as the characters wouldnt see the ghost, but i have it as place holder till i understand how to tell who is being possessed.