lots of things
added music, some sounds, AI and nav mesh changes. Messed with lighting to get a better understanding of it, etc. etc.

Still need to adjust and add sounds, door and key functionality, little better AI, rat pathing when the level is worked on, adjusting rain when possesed or not, etc etc.