changes to the AI
Made 2 public booleans that determines what the AI does when the player is in ghost form.
One determines if the AI stops when the player gets close if in ghost form. The AI will stop if you get close to it with a character other than the spirit.
(This is on by default because I figured it would be annoying to the player if they want to possess something and the AI are running away from them, but its toggle-able. Maybe best for rats)

The other determines if the AI will look at you or not in ghost form. They will look at you if you are in any character other than the spirit.
(This is off by default because I figured that it would make sense for the AI to not look at you when in spirit form.)