UI overhaul should be working
However implementation is not complete. We will need to populate every clue / item / character with the proper item UI text and associated icon. (generally the button to interact is on the top line, requirements / naming of object is bottom line)
Implementation for doors, reading, safe, etc, are NOT in just yet. Need to extend their current code from Item Abstract class, and set up their activation and UI behaviors.
(workin on it, will be soon, hopefully done today b4 meeting)